Move codepage-fixer to utils-module
