+++ src/zinolib/ritz.py
@@ -89,7 +89,10 @@ from typing import NamedTuple
 import codecs
 import select
 
+from .utils import windows_codepage_cp1252
 
+
+codecs.register_error("windows_codepage_cp1252", windows_codepage_cp1252)
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.ERROR)
 
@@ -107,52 +110,6 @@ class ProtocolError(Exception):
 
 NotifierResponse = NamedTuple("NotifierResponse", [("id", int), ("type", str), ("info", str)])
 DataResponse = NamedTuple("DataResponse", [("data", str), ("header", str)])
-
-
-def windows_codepage_cp1252(error):
-    """Windows Codepage 1252 decoder fallback
-
-    This function could be used as a failback for UnicodeEncodeError to fail
-    back to windows codepage 1252 decoding when eg. utf-8 failes
-
-    All valid windows codepage 1252 elements will be returned with their
-    representative unicode characters and failes back to unicode unknown char
-    when no valid character is found
-
-    All illegal characters are replaced with unicode 0xFFFD
-    """
-
-    # This only works on UnicodeDecodeError, will not work on encoding
-    if not isinstance(error, UnicodeDecodeError):
-        raise error
-
-    # 0xFFFD is the unicode char for illegal character.
-    # these characters are not an valid printable cp1252 character
-
-    cp1252_map = [
-        0x20AC, 0xFFFD, 0x201A, 0x0192, 0x201E, 0x2026, 0x2020, 0x2021, #0x80-0x87
-        0x02c6, 0x2030, 0x0160, 0x2039, 0x0152, 0xFFFD, 0x017d, 0xFFFD, #0x88-0x8F
-        0xFFFD, 0x2018, 0x2019, 0x201C, 0x201D, 0x2022, 0x2013, 0x2014, #0x90-0x97
-        0x02DC, 0x2122, 0x0161, 0x203A, 0x0153, 0xFFFD, 0x017E, 0x0178  #0x98-0x9F
-    ]  # fmt: skip
-
-    result = []
-    for i in range(error.start, error.end):
-        byte = error.object[i]
-
-        if byte >= 0x80 and byte <= 0x9F:
-            # We try to use windows codepage 1252 on this char
-            result.append(chr(cp1252_map[byte - 0x80]))
-
-        else:
-            # This looks like a valid latin-1 char
-            # It correspons to the same unicode char
-            result.append(chr(0x00 + byte))
-
-    return "".join(result), error.end
-
-
-codecs.register_error("windows_codepage_cp1252", windows_codepage_cp1252)
 
 
 class caseState(enum.Enum):
